voice and stickers

+++ main.go
@@ -17,10 +17,6 @@ type Handler struct {
 	db      *sql.DB
 }
 
-/*ype D_message struct{
-	message int
-}*/
-
 // предустановки
 const (
 	stop_find_button   = "❌ Остановить поиск"
@@ -31,10 +27,11 @@ const (
 	Registration_fine  = "✅ Регистрация успешно пройдена\nНажми на /start"
 	time_to_find       = "👽 Пора искать тебе собеседника"
 	person_find        = "Собеседник найден\nВся переписка защищена 🔒"
-	queue              = "❗Вы уже находитесь в поиске!"
-	stop_chat          = "Вы покинули чат"
+	queue              = "❗ Вы уже находитесь в поиске!"
+	stop_chat          = "🚶 Вы покинули чат"
 	leave_chat         = "😓 Собеседник покинул чат"
 	no_available_chat  = "🚫 У вас не было активного чата"
+	no_access 	       = "⚠ Пройдите регистрацию"
 )
 
 var sex_menu = tgbotapi.NewInlineKeyboardMarkup(
@@ -91,33 +88,34 @@ func (h Handler) Commands() {
 		}
 
 	case "stop":
-		if chat := h.db.Get_active_chat(h.updater.Message.Chat.ID); chat != (sql.Chat{}) {
-			h.db.Delete_chat(h.updater.Message.Chat.ID)
-			msg.ReplyMarkup = Start_menu
-			msg.Text = stop_chat
-			if _, err := h.bot.Send(msg); err != nil {
-				log.Panic(err)
-			}
-			msg.Text = leave_chat
-			if chat_one := int64(chat.Chat_one); chat_one != h.updater.Message.Chat.ID {
-				msg.ChatID = chat_one
-			} else if chat_two := int64(chat.Chat_two); chat_two != h.updater.Message.Chat.ID {
-				msg.ChatID = chat_two
-			}
+		if h.db.Check_person(h.updater.Message.From.UserName) == true {
 
-		} else if h.r.Queue_exist(h.updater.Message.Chat.ID) == true {
-			h.r.Queue_rem(h.updater.Message.Chat.ID)
-			msg.Text = stop_search_perosn
-			msg.ReplyMarkup = Start_menu
-		} else {
-			msg.Text = no_available_chat
+			if chat := h.db.Get_active_chat(h.updater.Message.Chat.ID); chat != (sql.Chat{}) {
+				h.db.Delete_chat(h.updater.Message.Chat.ID)
+				msg.ReplyMarkup = Start_menu
+				msg.Text = stop_chat
+				h.Send(msg)
+				msg.Text = leave_chat
+				if chat_one := int64(chat.Chat_one); chat_one != h.updater.Message.Chat.ID {
+					msg.ChatID = chat_one
+				} else if chat_two := int64(chat.Chat_two); chat_two != h.updater.Message.Chat.ID {
+					msg.ChatID = chat_two
+				}
+
+			} else if h.r.Queue_exist(h.updater.Message.Chat.ID) == true {
+				h.r.Queue_rem(h.updater.Message.Chat.ID)
+				msg.Text = stop_search_perosn
+				msg.ReplyMarkup = Start_menu
+			} else {
+				msg.Text = no_available_chat
+			}
+		}else{
+			msg.Text = no_access
 		}
 	default:
 		msg.Text = "Неизвестная команда"
 	}
-	if _, err := h.bot.Send(msg); err != nil {
-		log.Panic(err)
-	}
+	h.Send(msg)
 }
 
 // обработка сообщений
@@ -137,15 +135,15 @@ func (h Handler) Messages() {
 				msg.Text = person_find
 				msg.ReplyMarkup = tgbotapi.NewRemoveKeyboard(true)
 				fmt.Println(chat)
-				if _, err := h.bot.Send(msg); err != nil {
-					log.Panic(err)
-				}
+				h.Send(msg)
 				msg.ChatID = int64(chat)
+				h.Send(msg)
 
 			} else {
 				h.r.Queue_add(h.updater.Message.Chat.ID)
 				msg.ReplyMarkup = stop_menu
 				msg.Text = search_person
+				h.Send(msg)
 			}
 
 		} else {
@@ -156,36 +154,61 @@ func (h Handler) Messages() {
 		h.r.Queue_rem(h.updater.Message.Chat.ID)
 		msg.ReplyMarkup = Start_menu
 		msg.Text = stop_search_perosn
-
+		h.Send(msg)
 	case profile:
 		msg.Text = fmt.Sprintf("Имя: %s\n", h.updater.Message.From.FirstName)
 		msg.ReplyMarkup = Start_menu
+		h.Send(msg)
 
 	default:
-		msg.Text = "Неизвестная команда"
+		if chat := h.db.Get_active_chat(h.updater.Message.Chat.ID); chat != (sql.Chat{}) {
+			if h.updater.Message.Voice != nil {
+				voice := tgbotapi.NewVoice(h.updater.Message.Chat.ID, nil)
+				voice.File = tgbotapi.FileID(h.updater.Message.Voice.FileID)
+
+				if chat_one := int64(chat.Chat_one); chat_one != h.updater.Message.Chat.ID {
+					voice.ChatID = chat_one
+				} else if chat_two := int64(chat.Chat_two); chat_two != h.updater.Message.Chat.ID {
+					voice.ChatID = chat_two
+				}
+				h.Send(voice)
+
+			} else if h.updater.Message.Text != "" {
+				msg.Text = h.updater.Message.Text
+				if chat_one := int64(chat.Chat_one); chat_one != h.updater.Message.Chat.ID {
+					msg.ChatID = chat_one
+				} else if chat_two := int64(chat.Chat_two); chat_two != h.updater.Message.Chat.ID {
+					msg.ChatID = chat_two
+				}
+				h.Send(msg)
+			} else if h.updater.Message.Sticker != nil{
+				sticker := tgbotapi.NewSticker(h.updater.Message.Chat.ID, tgbotapi.FileID(h.updater.Message.Sticker.FileID))
+				if chat_one := int64(chat.Chat_one); chat_one != h.updater.Message.Chat.ID {
+					sticker.ChatID = chat_one
+				} else if chat_two := int64(chat.Chat_two); chat_two != h.updater.Message.Chat.ID {
+					sticker.ChatID = chat_two
+				}
+				h.Send(sticker)
+			}
+
+		} else {
+			msg.Text = "Неизвестная команда"
+			h.Send(msg)
+		}
 
 	}
+
+}
+
+func (h Handler) Send(msg tgbotapi.Chattable) {
 	if _, err := h.bot.Send(msg); err != nil {
 		log.Panic(err)
 	}
-
 }
 
-func (h Handler) Leave_chat(db *sql.DB) {
-
-}
 func (h Handler) Edit_message(chat_id int64, lastMessageID int) {
 	edit_message := tgbotapi.NewEditMessageText(chat_id, lastMessageID, Registration_fine)
-	if _, err := h.bot.Send(edit_message); err != nil {
-		log.Panic(err)
-	}
-}
-
-func (h Handler) Delete_chat(chat int64) {
-
-}
-func (h Handler) Callback(db *sql.DB) {
-	// здесь будут calbacks
+	h.Send(edit_message)
 }
 
 func main() {
